% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/91_addCheckId.R
\name{addCheckId}
\alias{addCheckId}
\title{addCheckId Add a checkId to the first column in a dataset}
\usage{
addCheckId(.ds, .checkId)
}
\arguments{
\item{.ds}{The dataset to add the column to}

\item{.checkId}{The checkId to add to the dataset}
}
\value{
The original dataset with .checkId added as the first column
}
\description{
addCheckId Add a checkId to the first column in a dataset
}
