% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/92_addSubsetDate.R
\name{addExcelReportSubsetDate}
\alias{addExcelReportSubsetDate}
\title{addExcelReportSubsetDate Add a subset date to the last column in a dataset called report_subset_date}
\usage{
addExcelReportSubsetDate(.ds, .subsetDateVariable)
}
\arguments{
\item{.ds}{The dataset to add the column to}

\item{.subsetDateVariable}{The checkId to add to the dataset}
}
\value{
The original dataset with .checkId added as the first column
}
\description{
addExcelReportSubsetDate Add a subset date to the last column in a dataset called report_subset_date
}
