% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_calculatePercentChange.R
\name{calculatePercentChange}
\alias{calculatePercentChange}
\title{calculatePercentChange Function to add a \% change column to a gtsummary object for NC24}
\usage{
calculatePercentChange(.data)
}
\arguments{
\item{.data}{A gtsummary object with 2 levels of a by variable}
}
\value{
The gtsummary object with an added column showing \% change
}
\description{
calculatePercentChange Function to add a \% change column to a gtsummary object for NC24
}
