% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_check_for_dq_reports.R
\name{check_for_dq_reports}
\alias{check_for_dq_reports}
\title{check_for_dq_reports}
\usage{
check_for_dq_reports(
  base_report_url,
  base_html_report_url,
  output_url,
  year_month = NULL
)
}
\arguments{
\item{base_report_url}{The url string to the base DQ reports folder}

\item{base_html_report_url}{The url string for the top level Biostat Data Files folder}

\item{output_url}{The url string for where I want to save the reports}

\item{year_month}{Optional; Specific month-year to look up the status; defaults to current month; enter as "YYYY-MM"}
}
\value{
A .csv table recording which registries have completed DQ reports for the current month, saved in the specified output location
}
\description{
Function to check for presence of DQ reports, both excel and html. Runs best when your working directory is registrydqchecks/R/
}
\note{
See code in registrydqchecks repository - ./R/99_check_for_dq_reports_test/04_check_for_dq_reports_example.R
}
