% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{check_missing_variables}
\alias{check_missing_variables}
\title{check_missing_variables Check that the expected variables exist in the given sheet}
\usage{
check_missing_variables(sheet_data, expectedVariables)
}
\arguments{
\item{sheet_data}{The data from the specific sheet in the Excel file}

\item{expectedVariables}{Character vector of variable names that should exist in the sheet}
}
\description{
check_missing_variables Check that the expected variables exist in the given sheet
}
