% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{check_missingness_threshold_multiplier}
\alias{check_missingness_threshold_multiplier}
\title{check_missingness_threshold_multiplier Check that missingnessThresholdMultiplier is a non-negative number (or missing)}
\usage{
check_missingness_threshold_multiplier(sheet_data)
}
\arguments{
\item{sheet_data}{Data from the specific sheet in the codebook}
}
\description{
check_missingness_threshold_multiplier Check that missingnessThresholdMultiplier is a non-negative number (or missing)
}
