% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{check_numeric_range}
\alias{check_numeric_range}
\title{check_numeric_range Check that numeric values fall within a specified range}
\usage{
check_numeric_range(sheet_data, var_name, min_val, max_val, condition_desc)
}
\arguments{
\item{sheet_data}{Data from the Excel sheet}

\item{var_name}{Name of the variable to check}

\item{min_val}{Minimum value of that variable}

\item{max_val}{Maximum value of that variable}

\item{condition_desc}{Description to print to the log if check does not pass}
}
\description{
check_numeric_range Check that numeric values fall within a specified range
}
