% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{check_numRange}
\alias{check_numRange}
\title{check_numRange Check the format and validity of numRange}
\usage{
check_numRange(sheet_data)
}
\arguments{
\item{sheet_data}{Data from the specific sheet in the codebook}
}
\description{
check_numRange Check the format and validity of numRange
}
