% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{check_variable_conditions}
\alias{check_variable_conditions}
\title{check_variable_conditions Function to check that specific variables fit specific given conditions}
\usage{
check_variable_conditions(sheet_data, var_name, valid_values, condition_desc)
}
\arguments{
\item{sheet_data}{Data from the specific sheet in the Excel file}

\item{var_name}{The variable name to check}

\item{valid_values}{The valid values of the given variable}

\item{condition_desc}{Description of the condition to print to the log}
}
\description{
check_variable_conditions Function to check that specific variables fit specific given conditions
}
