% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25_checkForAddedRows.R
\name{checkForAddedRows}
\alias{checkForAddedRows}
\title{(internal function) Check for how many rows were added to the dataset compared to last month}
\usage{
checkForAddedRows(.dsToCheck, .compDsToCheck)
}
\arguments{
\item{.dsToCheck}{A dataframe to check}

\item{.compDsToCheck}{A dataframe from last month}
}
\value{
A list with pass/fail, number of rows added, number of old rows, and the proportion increase in rows
}
\description{
(internal function) Check for how many rows were added to the dataset compared to last month
}
