% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_checkForDuplicateUniqueIds.R
\name{checkForDuplicateUniqueIds}
\alias{checkForDuplicateUniqueIds}
\title{(internal function) Check for duplicate unique IDs in a given dataset}
\usage{
checkForDuplicateUniqueIds(.dsToCheck, .uniqueKey)
}
\arguments{
\item{.dsToCheck}{A dataframe of the dataset to check}

\item{.uniqueKey}{A character vector that contains the names of the unique keys}
}
\value{
A list with pass/fail, number of duplicate rows, and a list of ids that are duplicates
}
\description{
(internal function) Check for duplicate unique IDs in a given dataset
}
