% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_checkForExtraVariables.R
\name{checkForExtraVariables}
\alias{checkForExtraVariables}
\title{Check that there are no extra variables compared to the variables that should be there}
\usage{
checkForExtraVariables(.dsToCheck, .listOfSupposedVars)
}
\arguments{
\item{.dsToCheck}{A dataframe of the dataset to check}

\item{.listOfSupposedVars}{A character vector with the names of the variables that should be in the dataset}
}
\value{
A list with pass/fail, the number of extra variables, and a list of the extra variables
}
\description{
Check that there are no extra variables compared to the variables that should be there
}
