% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27_checkForGivenItemsNonresponse.R
\name{checkForGivenItemsNonresponse}
\alias{checkForGivenItemsNonresponse}
\title{(internal function) Check the missingness for the given items}
\usage{
checkForGivenItemsNonresponse(.dsToCheck, .listOfEssentialVars)
}
\arguments{
\item{.dsToCheck}{A dataframe to check}

\item{.listOfEssentialVars}{A character vector with the variables deemed "essential"}
}
\value{
A list with pass/fail, and a dataframe with the information on the missingness for those variables
}
\description{
(internal function) Check the missingness for the given items
}
