% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/28_checkForMonthlyMissingness.R
\name{checkForMonthlyMissingness}
\alias{checkForMonthlyMissingness}
\title{(internal function) Check the month-to-month missingness for the given items}
\usage{
checkForMonthlyMissingness(.dsToCheck, .compDsToCheck, .listOfEssentialVars)
}
\arguments{
\item{.dsToCheck}{A dataframe to check}

\item{.compDsToCheck}{A dataframe to check from last month}

\item{.listOfEssentialVars}{A character vector with the variables deemed "essential"}
}
\value{
A list with pass/fail, and a dataframe with the information on the missingness for those variables between the two datasets
}
\description{
(internal function) Check the month-to-month missingness for the given items
}
