% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_checkForOmittedVariables.R
\name{checkForOmittedVariables}
\alias{checkForOmittedVariables}
\title{(internal function) Check for variables that have been left out of the comparator list}
\usage{
checkForOmittedVariables(.dsToCheck, .listOfSupposedVars)
}
\arguments{
\item{.dsToCheck}{A dataframe of the dataset to check}

\item{.listOfSupposedVars}{A character vector containing the list of variables that should be in the dataset}
}
\value{
A list with pass/fail, the number of omitted variables, and a list of the omitted variables
}
\description{
(internal function) Check for variables that have been left out of the comparator list
}
