% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26_checkForRemovedRows.R
\name{checkForRemovedRows}
\alias{checkForRemovedRows}
\title{(internal function) Check for rows that were in last months dataset but are not in this month}
\usage{
checkForRemovedRows(.dsToCheck, .compDsToCheck, .uniqueKey)
}
\arguments{
\item{.dsToCheck}{A dataframe to check}

\item{.compDsToCheck}{A dataframe from last month}

\item{.uniqueKey}{A character vector with names of variables that define the unique keys}
}
\value{
A list with pass/fail, number of removed rows, and a listing of unique keys that were removed
}
\description{
(internal function) Check for rows that were in last months dataset but are not in this month
}
