% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29a_checkForValidAgeAtEnrollment.R
\name{checkForValidAgeAtEnrollment}
\alias{checkForValidAgeAtEnrollment}
\title{checkForValidAgeAtEnrollment DQ check for CC10 - Check that age}
\usage{
checkForValidAgeAtEnrollment(
  .dsName,
  .dsToCheck,
  .codebookVariables,
  .uniqueKeys
)
}
\arguments{
\item{.dsName}{The name of the dataset being checked}

\item{.dsToCheck}{The dataset to check}

\item{.codebookVariables}{The codebook variables for the dataset}

\item{.uniqueKeys}{The list of unique keys}
}
\value{
The results of the check
}
\description{
checkForValidAgeAtEnrollment DQ check for CC10 - Check that age
}
