% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29_checkForVariableTypes.R
\name{checkForVariableTypes}
\alias{checkForVariableTypes}
\title{(internal function) Check if any variables have unexpected types}
\usage{
checkForVariableTypes(.dsToCheck, .codebookVariables)
}
\arguments{
\item{.dsToCheck}{A dataframe to check}

\item{.codebookVariables}{The codebook variables for this specific check}
}
\value{
A list with pass/fail, the number of variables with unexpected types, and a list of the variables with unexpected types
}
\description{
(internal function) Check if any variables have unexpected types
}
