% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/84_cleanUniqueKeyClasses.R
\name{cleanUniqueKeyClasses}
\alias{cleanUniqueKeyClasses}
\title{cleanUniqueKeyClasses Remove vctrs_vctr and haven_labelled from the classes of uniqueKey variables}
\usage{
cleanUniqueKeyClasses(datasetToCheck, uniqueKeyVars)
}
\arguments{
\item{data}{The dataset to remove vctrs_vctr and haven_labelled from the classes of uniqueKey variables}
}
\value{
the dataset with vctrs_vctr and haven_labelled removed from the classes of the uniqueKey variables
}
\description{
cleanUniqueKeyClasses Remove vctrs_vctr and haven_labelled from the classes of uniqueKey variables
}
