% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_codebookNcChecks.R
\name{codebookNcChecks}
\alias{codebookNcChecks}
\title{codebookNcChecks Run the codebook-specific noncritical checks and return the values}
\usage{
codebookNcChecks(.dsName, .dsToCheck, .compDsToCheck, .codebookVariables)
}
\arguments{
\item{.dsName}{Name of the dataset being checked}

\item{.dsToCheck}{The dataset being checked}

\item{.compDsToCheck}{The comparator dataset being checked}

\item{.codebookVariables}{The variables from the codebook to run the checks}
}
\value{
.resultsOfCodebookNcChecks results of the codebookNcChecks
}
\description{
codebookNcChecks Run the codebook-specific noncritical checks and return the values
}
