% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_copyDataset.R
\name{confirmDeletion}
\alias{confirmDeletion}
\title{confirmDeletion Requires manual typed confirmation of file deletion}
\usage{
confirmDeletion(.dsName)
}
\arguments{
\item{.dsName}{The name of the dataset to be deleted}
}
\value{
Boolean - TRUE if confirmed, FALSE if not
}
\description{
confirmDeletion Requires manual typed confirmation of file deletion
}
