% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{convert_to_numeric Convert columns from character to numeric}
\usage{
convert_to_numeric(column)
}
\arguments{
\item{column}{Name of the column to convert}
}
\description{
convert_to_numeric Convert columns from character to numeric
}
