% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_copyRomListingToFolder.R
\name{copyRomListingToFolder}
\alias{copyRomListingToFolder}
\title{copyRomListingToFolder Copy the CDM/ROM listing to the appropriate folder upon request}
\usage{
copyRomListingToFolder(
  .reportOutputUrl,
  .romReportUrl,
  .registry,
  .dataPullDate,
  .overwrite = FALSE
)
}
\arguments{
\item{.reportOutputUrl}{The url to the base report folder where the DQ report has been output}

\item{.romReportUrl}{The url to the location for ROM/CDM to view the listing}

\item{.registry}{A text string with the initials of the specific registry ("ad", "ms", "raj", "aa", "pso", "ibd", "nmo")}

\item{.dataPullDate}{A date string of the data pull date in the format YYYY-DD-MM (e.g. "2024-01-10")}

\item{.overwrite}{Whether or not to overwrite the currently existing report in the CDM/ROM folder (DEFAULT = FALSE)}
}
\value{
The full string URL to the location of the Excel file in the new location
}
\description{
copyRomListingToFolder Copy the CDM/ROM listing to the appropriate folder upon request
}
