% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51_createDataStoreFolder.R
\name{createDataStoreFolder}
\alias{createDataStoreFolder}
\title{(internal function) Create a new folder at the specified location if it doesn't already exist}
\usage{
createDataStoreFolder(.urlToCreate)
}
\arguments{
\item{.urlToCreate}{A url string to the folder to be created}
}
\description{
(internal function) Create a new folder at the specified location if it doesn't already exist
}
