% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_changes.R
\name{data_changes}
\alias{data_changes}
\title{Data changes function}
\usage{
data_changes(
  curr_dataset,
  comp_dataset,
  by_vars,
  output_folder,
  output_filename,
  title_pagename
)
}
\arguments{
\item{curr_dataset}{the current data set to be compared}

\item{comp_dataset}{the data set to compare to the current data set}

\item{by_vars}{list of by variables denoting a unique record in the data sets}

\item{output_folder}{the output folder path}

\item{output_filename}{the output file name}

\item{title_pagename}{the title page name for the title page prefixed before ": Data Changes" (ie. AA: Data Changes)}
}
\value{
an output excel report in the specified output folder and named as the output file name
}
\description{
Function to run comparison between the current data set and the provided comparison data set
}
