% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{dataset_exists}
\alias{dataset_exists}
\title{dataset_exists Check if a given dataset name has a corresponding sheet in the codebook}
\usage{
dataset_exists(dataset, codebookSheets)
}
\arguments{
\item{dataset}{The dataset name to check against the sheet names}

\item{codebookSheets}{Names of the sheets in the codebook}
}
\description{
dataset_exists Check if a given dataset name has a corresponding sheet in the codebook
}
