% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_runCategoricalValueChecks.R
\name{extractValueLabels}
\alias{extractValueLabels}
\title{extractValueLabels Extract the values after the equal sign in each element in a vector}
\usage{
extractValueLabels(vec)
}
\arguments{
\item{vec}{A vector to extract labels from}
}
\value{
A vector of the labels of the levels of the variable
}
\description{
extractValueLabels Extract the values after the equal sign in each element in a vector
}
