% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_runCategoricalValueChecks.R
\name{extractValueNumbers}
\alias{extractValueNumbers}
\title{extractValueNumbers Extract the numbers before the equal sign in each element in a vector}
\usage{
extractValueNumbers(vec)
}
\arguments{
\item{vec}{A vector to extract numbers from}
}
\value{
A vector of numbers of the levels of the variable
}
\description{
extractValueNumbers Extract the numbers before the equal sign in each element in a vector
}
