% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_calculatePercentChange.R
\name{filterPercentChange}
\alias{filterPercentChange}
\title{filterPercentChange Filter a gtsummary object to only show rows with percentChange >5}
\usage{
filterPercentChange(data)
}
\arguments{
\item{data}{The gtsummary object to filter}
}
\value{
The filtered dataset
}
\description{
filterPercentChange Filter a gtsummary object to only show rows with percentChange >5
}
