% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53_generateCheckSummary.R
\name{generateCheckSummary}
\alias{generateCheckSummary}
\title{generateCheckSummary}
\usage{
generateCheckSummary(.resultsOfChecks)
}
\arguments{
\item{.resultsOfChecks}{A list containing "criticalChecks" and "nonCriticalChecks" in the appropriate CE DQ format}
}
\value{
A list with the summaries of the critical and noncritical checks
}
\description{
generateCheckSummary
}
