% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/54_generateCriticalCheckSummary.R
\name{generateCriticalCheckSummary}
\alias{generateCriticalCheckSummary}
\title{generateCriticalCheckSummary}
\usage{
generateCriticalCheckSummary(.criticalChecksToSummarize)
}
\arguments{
\item{.criticalChecksToSummarize}{The critical checks to be summarized in the appropriate CE DQ format}
}
\value{
Dataframe with a summary of the critical checks for the datasets
}
\description{
generateCriticalCheckSummary
}
