% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55_generateNoncriticalCheckSummary.R
\name{generateNoncriticalCheckSummary}
\alias{generateNoncriticalCheckSummary}
\title{generateNoncriticalCheckSummary}
\usage{
generateNoncriticalCheckSummary(.nonCriticalChecksToSummarize)
}
\arguments{
\item{.nonCriticalChecksToSummarize}{The noncritical checks to be summarized in the appropriate CE DQ format}
}
\value{
Dataframe with a summary of the critical checks for the datasets
}
\description{
generateNoncriticalCheckSummary
}
