% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52_outputListings.R
\name{outputListings}
\alias{outputListings}
\title{Output specified listings in xlsx format for further review}
\usage{
outputListings(
  .registry,
  .listingUrl,
  .yearMonthTimestamp,
  .dataPullDate,
  .timestamp,
  .checksToOutput,
  .activeSites
)
}
\arguments{
\item{.registry}{Name of the registry being checked}

\item{.listingUrl}{Url to where the listing output will live}

\item{.yearMonthTimestamp}{The year and month of the current datapull}

\item{.dataPullDate}{The YYYY-MM-DD of the current datapull}

\item{.timestamp}{Timestamp of the current run}

\item{.checksToOutput}{Registry DQ checks to output in list form with criticalChecks and nonCriticalChecks entries}

\item{.activeSites}{Information on the list of active sites for this registry}
}
\value{
A text string with the URL to the CDM/ROM Excel listing
}
\description{
Output specified listings in xlsx format for further review
}
