% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56_perpetuateExcelComments.R
\name{perpetuateExcelComments}
\alias{perpetuateExcelComments}
\title{perpetuateExcelComments A function to read in the CDM/ROM report from last month and pull comments and place them in the new report}
\usage{
perpetuateExcelComments(
  .lastMonthCheckExcelFileUrl,
  .thisMonthCheckExcelFileUrl
)
}
\arguments{
\item{.lastMonthCheckExcelFileUrl}{A direct URL to the Excel file from last month}

\item{.thisMonthCheckExcelFileUrl}{A direct URL to the Excel file from this month}
}
\description{
perpetuateExcelComments A function to read in the CDM/ROM report from last month and pull comments and place them in the new report
}
