% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{process_dataset}
\alias{process_dataset}
\title{process_dataset A function to validate each tab of the codebook against the specifications}
\usage{
process_dataset(codebookUrl, dataset, expectedVariables)
}
\arguments{
\item{codebookUrl}{Url to the codebook of interest}

\item{dataset}{Prefix of the dataset to check in the codebook}

\item{expectedVariables}{List of variables expected to be in the codebook}
}
\description{
process_dataset A function to validate each tab of the codebook against the specifications
}
