% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_pullCodebookFromExcelFile.R
\name{pullCodebookFromExcelFile}
\alias{pullCodebookFromExcelFile}
\title{(internal function) Pull a registrydqchecks specific codebook from a specified location}
\usage{
pullCodebookFromExcelFile(.fileUrl, .sheetName)
}
\arguments{
\item{.fileUrl}{A text url to the Excel file to pull in}

\item{.sheetName}{A text string with the name of the sheet to pull from}
}
\value{
A dataframe with the columns from the file that had been specified
}
\description{
(internal function) Pull a registrydqchecks specific codebook from a specified location
}
