% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_pullData.R
\name{pullData}
\alias{pullData}
\title{(internal function) A function to pull the R dataset from the given url location}
\usage{
pullData(.datasetUrl, .isR = TRUE)
}
\arguments{
\item{.datasetUrl}{A url string that points to the location of the dataset}

\item{.isR}{A boolean indicating if the dataset is a .RDS or a .DTA (default = TRUE)}
}
\value{
A dataframe of the dataset that was pulled
}
\description{
(internal function) A function to pull the R dataset from the given url location
}
