% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_pullDTAfromUrl.R
\name{pullDTAfromUrl}
\alias{pullDTAfromUrl}
\title{(internal function) Pull a .DTA file from a specified path}
\usage{
pullDTAfromUrl(.datasetUrl)
}
\arguments{
\item{.datasetUrl}{A url string to the .DTA file being pulled}
}
\value{
A dataframe of the file being pulled
}
\description{
(internal function) Pull a .DTA file from a specified path
}
