% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_pullRDSfromUrl.R
\name{pullRDSfromUrl}
\alias{pullRDSfromUrl}
\title{(internal function) Pull a .RDS file from a specified path}
\usage{
pullRDSfromUrl(.datasetUrl)
}
\arguments{
\item{.datasetUrl}{A url string to the .RDS file being pulled}
}
\value{
A dataframe of the file being pulled
}
\description{
(internal function) Pull a .RDS file from a specified path
}
