% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/87_pullSiteInfoFromExcelFile.R
\name{pullSiteInfoFromExcelFile}
\alias{pullSiteInfoFromExcelFile}
\title{(internal function) pullSiteInfoFromExcelFile Pull a site information}
\usage{
pullSiteInfoFromExcelFile(.fileUrl, .registry)
}
\arguments{
\item{.fileUrl}{A text url to the Excel file to pull in}

\item{.registry}{The acronym for the current registry}
}
\value{
A dataframe with the columns from the file that had been specified
}
\description{
(internal function) pullSiteInfoFromExcelFile Pull a site information
}
