% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/81_pullVariableListFromExcelFile.R
\name{pullVariableListFromExcelFile}
\alias{pullVariableListFromExcelFile}
\title{(internal function) Pull a list of variables from a specific column in a specific excel file}
\usage{
pullVariableListFromExcelFile(.fileUrl, .fileName, .sheetName, .colName)
}
\arguments{
\item{.fileUrl}{A text url to the location of the Excel file}

\item{.fileName}{A text string with the name of the file}

\item{.sheetName}{A text string with the name of the sheet to pull from}

\item{.colName}{A character vector with the names of the columns to pull from}
}
\value{
A dataframe that contains the columns that have been specified
}
\description{
(internal function) Pull a list of variables from a specific column in a specific excel file
}
