% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{read_dataset}
\alias{read_dataset}
\title{read_dataset Read the data from a specific sheet in the codebook}
\usage{
read_dataset(codebookUrl, dataset)
}
\arguments{
\item{codebookUrl}{A url to the codebook}

\item{dataset}{The name of the dataset tab}
}
\description{
read_dataset Read the data from a specific sheet in the codebook
}
