% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{read_excel_sheets Read the names of the sheets in the given excel file}
\usage{
read_excel_sheets(codebookUrl)
}
\arguments{
\item{codebookUrl}{A url to the codebook}
}
\description{
read_excel_sheets Read the names of the sheets in the given excel file
}
