% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_runCategoricalValueChecks.R
\name{removeQuotes}
\alias{removeQuotes}
\title{removeQuotes A function to remove quotes from each element in a character string}
\usage{
removeQuotes(vec)
}
\arguments{
\item{vec}{A vector to remove quotes from}
}
\description{
removeQuotes A function to remove quotes from each element in a character string
}
