% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_runRangeAndValueChecks.R
\name{replaceNullWithNa}
\alias{replaceNullWithNa}
\title{replaceNullWithNa Replace NULL values with NA in specified columns of a dataframe}
\usage{
replaceNullWithNa(df, columns)
}
\arguments{
\item{df}{A dataframe in which to replace NULL values}

\item{columns}{A vector of column names or indices where NULL values should be replaced with NA}
}
\value{
The dataframe with NULL values replaced with NA in the specified columns
}
\description{
replaceNullWithNa Replace NULL values with NA in specified columns of a dataframe
}
