% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_runCategoricalValueChecks.R
\name{runCategoricalValueChecks}
\alias{runCategoricalValueChecks}
\title{runCategoricalValueChecks Run a check to ensure categorical variables have the appropriate values}
\usage{
runCategoricalValueChecks(.dsName, .dsToCheck, .codebookVariables, .uniqueKeys)
}
\value{
Results of the check
}
\description{
runCategoricalValueChecks Run a check to ensure categorical variables have the appropriate values
}
