% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_runNumericRangeChecks.R
\name{runNumericRangeChecks}
\alias{runNumericRangeChecks}
\title{runNumericRangeChecks Run a check to ensure numeric variables fall within an appropriate range}
\usage{
runNumericRangeChecks(.dsName, .dsToCheck, .codebookVariables, .uniqueKeys)
}
\arguments{
\item{.dsName}{The name of the dataset being checked}

\item{.dsToCheck}{The dataset being checked}

\item{.codebookVariables}{The codebook variabled needed for this specific check}

\item{.uniqueKeys}{The unique keys for the dataset being checked}
}
\value{
Results of the check
}
\description{
runNumericRangeChecks Run a check to ensure numeric variables fall within an appropriate range
}
