% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mainFunctions.R
\name{runRegistryChecks}
\alias{runRegistryChecks}
\title{The main function call to run the CorEvitas Registry DQ Checks and the Report}
\usage{
runRegistryChecks(
  .registry = "defaultRegistry",
  .prelimDataFolderUrl,
  .prelimDataPullDate,
  .lastMonthDataFolderUrl,
  .lastMonthDataPullDate,
  .codebookUrl,
  .siteInfoUrl = NULL,
  .cdmRomReportUrl,
  .datasetsToCheck,
  .nonCriticalChecks = NULL,
  .outputUrl,
  .isR
)
}
\arguments{
\item{.registry}{A text string with the initials of the specific registry ("ad", "ms", "raj", "aa", "pso", "ibd", "nmo")}

\item{.prelimDataFolderUrl}{(FOLDER URL) A url string to the FOLDER with the preliminary datasets}

\item{.prelimDataPullDate}{A date string of the data pull date in the format YYYY-DD-MM (e.g. "2024-01-10")}

\item{.lastMonthDataFolderUrl}{(FOLDER URL) A url string to the FOLDER with the last month's datasets}

\item{.lastMonthDataPullDate}{A date string of last month's data pull date in the format YYYY-DD-MM (e.g. "2023-12-05")}

\item{.codebookUrl}{(FILE URL - .XLSX) A url string to the registry-specific codebook FILE (MUST BE DIRECTLY TO THE .XLSX FILE)}

\item{.siteInfoUrl}{(FILE URL - .CSV) A Url to the FILE for site/provider data (MUST BE DIRECTLY TO THE site_data_clean.csv FILE)}

\item{.cdmRomReportUrl}{(FOLDER URL) A url string to the location of the base CDM/ROM Report folder}

\item{.datasetsToCheck}{A string vector with the names of the datasets to be checked (e.g. c("exvisit", "exlab", "exdrugexp")) - NOTE: These must perfectly match both the tab names in the codebook AND the names of the datasets being checked}

\item{.nonCriticalChecks}{A list of the manually generated non-critical checks in the CE DQ specified format (see additional documentation)}

\item{.outputUrl}{(FOLDER URL) A url string to the location of the output datasets - NOTE: A subfolder will be created here called /checks that will house the results of the checks and will be the location called by the check report}

\item{.isR}{A boolean indicating if the datasets being checked are in R or Stata format (e.g. if R then .isR = TRUE; if Stata then .isR = FALSE)}
}
\value{
A personalized URL to the location of the report
}
\description{
The main function call to run the CorEvitas Registry DQ Checks and the Report
}
