% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_checkDataStore.R
\name{submitToDataStore}
\alias{submitToDataStore}
\title{(internal function) Submit the results of the data quality checks to a set location}
\usage{
submitToDataStore(.registry, .dsPullDate, .dataStoreUrl, .resultsOfChecks)
}
\arguments{
\item{.registry}{A text string with the initials of the registry (e.g. "ad", "pso", "ms")}

\item{.dsPullDate}{A date string of the data pull date in YYYY-MM-DD format (e.g. "2024-01-10")}

\item{.dataStoreUrl}{A text url of the location where the check results will be stored}

\item{.resultsOfChecks}{A list with the results of the data checks}
}
\description{
(internal function) Submit the results of the data quality checks to a set location
}
