% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/86_subsetDatasetToActiveSites.R
\name{subsetDatasetToActiveSites}
\alias{subsetDatasetToActiveSites}
\title{subsetDatasetToActiveSites Function to subset dataset based on active sites}
\usage{
subsetDatasetToActiveSites(.dataset, .siteVar1, .siteVar2, .activeSites)
}
\arguments{
\item{.dataset}{The dataset to subset}

\item{.siteVar1}{The primary site variable to check for}

\item{.siteVar2}{The secondary site variable to check for}

\item{.activeSites}{The date of the current datapull}
}
\value{
The subset dataset
}
\description{
subsetDatasetToActiveSites Function to subset dataset based on active sites
}
