% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/85_subsetDatasetToLastYear.R
\name{subsetDatasetToLastYear}
\alias{subsetDatasetToLastYear}
\title{subsetDatasetToLastYear Function to subset datasetA based on the existence of variables}
\usage{
subsetDatasetToLastYear(.dataset, .timeVar1, .timeVar2, .dataPullDate)
}
\arguments{
\item{.dataset}{The dataset to subset}

\item{.timeVar1}{The primary time variable to check for}

\item{.timeVar2}{The secondary variable to check for}

\item{.dataPullDate}{The date of the current datapull}
}
\value{
The subset dataset
}
\description{
subsetDatasetToLastYear Function to subset datasetA based on the existence of variables
}
