% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88_subsetDatasetToNonCalculatedVariables.R
\name{subsetDatasetToNonCalculatedVariables}
\alias{subsetDatasetToNonCalculatedVariables}
\title{subsetDatasetToNonCalculatedVariables Function to subset dataset based on active sites}
\usage{
subsetDatasetToNonCalculatedVariables(.dataset)
}
\arguments{
\item{.dataset}{The dataset to subset}
}
\value{
The subset dataset
}
\description{
subsetDatasetToNonCalculatedVariables Function to subset dataset based on active sites
}
