% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mainFunctions.R
\name{updatePackages}
\alias{updatePackages}
\title{updatePackages}
\usage{
updatePackages(snapshot = TRUE)
}
\arguments{
\item{snapshot}{Boolean (default = TRUE) to automatically run renv::snapshot}
}
\description{
updatePackages
}
