% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/83_validateCodebook.R
\name{validateCodebook}
\alias{validateCodebook}
\title{validateCodebook Function to check structure of the codebook against specifications}
\usage{
validateCodebook(codebookUrl, datasetNames)
}
\arguments{
\item{codebookUrl}{Url to the codebook}

\item{datasetNames}{Character vector of dataset prefixes}
}
\description{
validateCodebook Function to check structure of the codebook against specifications
}
