% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_checkForMissingVariableLabels.R
\name{checkForMissingVariableLabels}
\alias{checkForMissingVariableLabels}
\title{Check if any variables have missing labels}
\usage{
checkForMissingVariableLabels(.dsToCheck)
}
\arguments{
\item{.dsToCheck}{A dataframe to check}
}
\value{
A list with pass/fail, the number of variables with missing labels, and a list of the variables with missing labels
}
\description{
Check if any variables have missing labels
}
