% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_close_sink.R
\name{close_sink}
\alias{close_sink}
\title{close sink file}
\usage{
close_sink(temp_log_file, registry_abbrev, output_location)
}
\arguments{
\item{temp_log_file}{The temporary log file to print out to the output location}

\item{registry_abbrev}{registry abbreviation (ie. "aa")}

\item{output_location}{report output location (ie. reportLocationUrl)}
}
\value{
saved console_output.txt file in the output location folder
}
\description{
close sink file
}
