% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_copyRomListingToFolder.R
\name{copyRomListingToFolder}
\alias{copyRomListingToFolder}
\title{copyRomListingToFolder Copy the CDM/ROM listing to the appropriate folder upon request}
\usage{
copyRomListingToFolder(.reportOutputUrl, .romReportUrl)
}
\arguments{
\item{.reportOutputUrl}{The url to the base report folder where the DQ report has been output}

\item{.romReportUrl}{The url to the location for ROM/CDM to view the listing}
}
\value{
The URL
}
\description{
copyRomListingToFolder Copy the CDM/ROM listing to the appropriate folder upon request
}
