% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_open_sink.R
\name{open_sink}
\alias{open_sink}
\title{open sink txt file to store console output in the console_output.txt}
\usage{
open_sink()
}
\value{
the temporary log file to be copied at the end of the script
}
\description{
open sink txt file to store console output in the console_output.txt
}
