% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52_outputListings.R
\name{outputListings}
\alias{outputListings}
\title{Output specified listings in xlsx format for further review}
\usage{
outputListings(.listingUrl, .timestamp, .checksToOutput)
}
\arguments{
\item{.listingUrl}{Url to where the listing output will live}

\item{.timestamp}{Timestamp of the current run}

\item{.checksToOutput}{Registry DQ checks to output}
}
\description{
Output specified listings in xlsx format for further review
}
