% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_parameterCheck.R
\name{paramCheck}
\alias{paramCheck}
\title{Parameter Checks}
\usage{
paramCheck(
  .registry,
  .prelimDataFolderUrl,
  .lastMonthDataFolderUrl,
  .codebookUrl,
  .siteInfoUrl,
  .cdmRomReportUrl,
  .outputUrl
)
}
\arguments{
\item{.registry}{}

\item{.prelimDataFolderUrl}{}

\item{.lastMonthDataFolderUrl}{}

\item{.codebookUrl}{}

\item{.siteInfoUrl}{}

\item{.cdmRomReportUrl}{}

\item{.outputUrl}{}
}
\value{
Print message that all parameters pass
}
\description{
Parameter Checks
}
\examples{

.registry = "ra"
.prelimDataFolderUrl = "C:/Users/olsonr/PPD (CRG)/Biostat Data Files - RA/DQ checks/Data/2025/2025-04-01/"
.lastMonthDataFolderUrl = "C:/Users/olsonr/PPD (CRG)/Biostat Data Files - RA/DQ checks/Data/2025/2025-03-01/"
.codebookUrl = "C:/Users/olsonr/PPD (CRG)/Biostat Data Files - RA/DQ checks/Documentations/RA_codebook_2025-04-01.xlsx"
.siteInfoUrl = "C:/Users/olsonr/PPD (CRG)/Biostat Data Files - Site and Provider Data/data/site_data_clean.csv"
.cdmRomReportUrl = "C:/Users/olsonr/PPD (CRG)/Core_Biostat and Epi Team Site - Documents/Biostat Registry Data Quality Reports"
.outputUrl = "C:/Users/olsonr/Documents/Scrap/ra/2025/2025-04/"

paramCheck(.registry = .registry
,.prelimDataFolderUrl = .prelimDataFolderUrl
,.lastMonthDataFolderUrl = .lastMonthDataFolderUrl
,.codebookUrl = .codebookUrl
,.siteInfoUrl = .siteInfoUrl
,.cdmRomReportUrl = .cdmRomReportUrl
,.outputUrl = .outputUrl)
}
