% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_runRangeAndValueChecks.R
\name{replaceNullWithNa}
\alias{replaceNullWithNa}
\title{replace_null_with_na Replace NULLs in dataset with NAs}
\usage{
replaceNullWithNa(df)
}
\arguments{
\item{df}{the dataset to transform}
}
\value{
the dataframe with NULLs transformed to NA
}
\description{
replace_null_with_na Replace NULLs in dataset with NAs
}
