% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_runRangeAndValueChecks.R
\name{runRangeAndValueChecks}
\alias{runRangeAndValueChecks}
\title{runRangeAndValueChecks Run nc1 which checks the numeric and categorical variables against their supposed values}
\usage{
runRangeAndValueChecks(.dsName, .dsToCheck, .codebookVariables, .uniqueKeys)
}
\value{
The results of the numeric range and categorical value checks
}
\description{
runRangeAndValueChecks Run nc1 which checks the numeric and categorical variables against their supposed values
}
