% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35_runReasonableMissingnessCheck.R
\name{runReasonableMissingnessCheck}
\alias{runReasonableMissingnessCheck}
\title{runReasonableMissingnessCheck Run a check to look at reasonable amount of missingness for all variables}
\usage{
runReasonableMissingnessCheck(.dsToCheck, .codebookVariables)
}
\arguments{
\item{.dsToCheck}{The dataset to run the checks on}

\item{.codebookVariables}{The codebook variables needed to run this specific check}
}
\value{
Results of the check
}
\description{
runReasonableMissingnessCheck Run a check to look at reasonable amount of missingness for all variables
}
