% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mainFunctions.R
\name{runRegistryChecks}
\alias{runRegistryChecks}
\title{The main function call to run the CorEvitas Registry DQ Checks and the Report}
\usage{
runRegistryChecks(
  .registry = "defaultRegistry",
  .prelimDataFolderUrl,
  .prelimDataPullDate,
  .lastMonthDataFolderUrl,
  .lastMonthDataPullDate,
  .codebookUrl,
  .datasetsToCheck,
  .outputUrl,
  .isR
)
}
\arguments{
\item{.registry}{A text string with the initials of the specific registry (e.g. "ad", "ms", "pso")}

\item{.prelimDataFolderUrl}{A url string to the folder with the preliminary datasets}

\item{.prelimDataPullDate}{A date string of the data pull date in the format YYYY-DD-MM (e.g. "2024-01-10")}

\item{.lastMonthDataFolderUrl}{A url string to the folder with the last month's datasets}

\item{.lastMonthDataPullDate}{A date string of last month's data pull date in the format YYYY-DD-MM (e.g. "2023-12-05")}

\item{.codebookUrl}{A url string to the registry-specific codebook}

\item{.datasetsToCheck}{A string vector with the names of the datasets to be checked (e.g. c("exvisit", "exlab", "exdrugexp")) - NOTE: These must perfectly match both the tab names in the codebook AND the names of the datasets being checked}

\item{.outputUrl}{A url string to the location of the output datasets - NOTE: A subfolder will be created here called /checks that will house the results of the checks and will be the location called by the check report}

\item{.isR}{A boolean indicating if the datasets being checked are in R or Stata format (e.g. if R then .isR = TRUE; if Stata then .isR = FALSE)}
}
\description{
The main function call to run the CorEvitas Registry DQ Checks and the Report
}
