% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_checkDataStore.R
\name{submitToDataStore}
\alias{submitToDataStore}
\title{(internal function) Submit the results of the data quality checks to a set location}
\usage{
submitToDataStore(
  .registry,
  .dsPullDate,
  .timestamp,
  .folderName,
  .dataStoreUrl,
  .resultsOfChecks,
  .activeSites,
  .dataFolderUrl,
  .lastMonthDataFolderUrl,
  .cdmRomReportUrl,
  .lastMonthDataPullDate
)
}
\arguments{
\item{.registry}{A text string with the initials of the registry (e.g. "ad", "pso", "ms")}

\item{.dsPullDate}{A date string of the data pull date in YYYY-MM-DD format (e.g. "2024-01-10")}

\item{.timestamp}{Timestamp that the checks were run}

\item{.folderName}{Name for the folder where results will be output}

\item{.dataStoreUrl}{A text url of the location where the check results will be stored}

\item{.resultsOfChecks}{A list with the results of the data checks}

\item{.activeSites}{Information on the list of active sites for this registry}

\item{.dataFolderUrl}{Folder URL of the current month dataset}

\item{.lastMonthDataFolderUrl}{Folder URL of the last month dataset}

\item{.cdmRomReportUrl}{(FOLDER URL) A url string to the location of the base CDM/ROM Report folder}

\item{.lastMonthDataPullDate}{A date string of last month's data pull date in the format YYYY-DD-MM (e.g. "2023-12-05")}
}
\description{
(internal function) Submit the results of the data quality checks to a set location
}
