% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_updatePackages.R
\name{updatePackages}
\alias{updatePackages}
\title{updatePackages Update rdc packages to the most recent version}
\usage{
updatePackages(snapshot = TRUE)
}
\arguments{
\item{snapshot}{Boolean (default = TRUE) to automatically run renv::snapshot}
}
\description{
updatePackages Update rdc packages to the most recent version
}
